requirements.txt
pytest
investpy
httpx
pre-commit
GitPython
click
fire